+++ admin/start/env.ts
@@ -29,6 +29,8 @@ export default await Env.create(new URL('../', import.meta.url), {
   NOMAD_STORAGE_PATH: Env.schema.string.optional(),
   NOMAD_BACKUP_PATH: Env.schema.string.optional(),
   NOMAD_RCLONE_REMOTE: Env.schema.string.optional(),
+  NOMAD_GUEST_MODE: Env.schema.boolean.optional(),
+  NOMAD_GUEST_TOOLS: Env.schema.string.optional(),
 
   /*
   |----------------------------------------------------------