+++ admin/start/kernel.ts
@@ -24,11 +24,12 @@ server.errorHandler(() => import('#exceptions/handler'))
  */
 server.use([
   () => import('#middleware/container_bindings_middleware'),
+  () => import('#middleware/guest_kiosk_middleware'),
   () => import('@adonisjs/cors/cors_middleware'),
   () => import('@adonisjs/vite/vite_middleware'),
   () => import('@adonisjs/inertia/inertia_middleware'),
   () => import('@adonisjs/static/static_middleware'),
-  () => import('#middleware/maps_static_middleware')
+  () => import('#middleware/maps_static_middleware'),
 ])
 
 /**